Added append(int), first(), last() and size()
append(int) - Adds a value at the end of the list and increases the size
by 50% if it is full.

first() - Returns the first element in the list or -1 if its empty.

last() - Returns the last element in the list or -1 if its empty.

size() - Returns the number of possible locations in the list whether or
not they have values assigned to them.